crates/sel4: Organize x86 modules
