+++ crates/sel4/src/arch/x86/arch/x64/mod.rs
@@ -4,15 +4,12 @@
 // SPDX-License-Identifier: MIT
 //
 
-mod invocations;
 mod object;
 mod user_context;
-mod vspace;
 
 pub(crate) mod top_level {
     pub use super::{
         object::{ObjectBlueprintSeL4Arch, ObjectBlueprintX64, ObjectTypeSeL4Arch, ObjectTypeX64},
         user_context::UserContext,
-        vspace::{FrameObjectType, TranslationStructureObjectType},
     };
 }
